Correct tests
Remove new files, can use existing ones
